Actually try to handle browser errors
The google u2f api has yet to have any error handling. The solution using an iframe is a terrible hack, as there is no way to know if postMessage fails. Instead, we now assume failure after a "reasonable" timeout, whatever that is. In this commit, I blindly set it to 200ms, without having a clue if this is orders of magnitude bad in either direction.

The main problem is caused by postMessage being broken by design, since you don't get success/failure information about the sending of the message.

Anyhow, the promises now actually do get rejected, if support for U2F seems to be lacking. This is tested with success on Opera and Chrome.