Use function expressions to avoid hoisting behavior
This commit changes `function cb` to `cb = function()` so that hosting
of functions does not change behavior.  At least with uglifyjs the
functions from both branches are hosted to the top of the function
under the same name which causes the callback from the non native
branch to be used in the native branch.