Ensures the check for an already existing backend when it fails
In our codebase, we have the need to retry `isSupported( )`
calls. This is due to, sometimes, the extension taking a while to
respond to the U2F Api. However, since the `_backend` is a promise
object, after the first call, it stays as resolved for the whole
lifecycle of the library.

This commit changes `_backend` variable to null when not supported
occurs, so the `getBackend( )` function forces the creation of a
new Promise. This allows the user to create a promise chaining
mechanism to retry the `isSupported( )` validation.